Add PostCodeDataVersion to the persisted data
This commit is exactly a cherry pick of the master commit
`20297dfbd96709af18b6046db76c99c79c149d59` which adds the
PostCodeDataVersion field. But the only change is the version is set to
0xFFFF (since the upstream is using 0 & 1 for decudicing absence of the
format version & to indicate of the format has been changed).

With this commit, any firmware update from 1060 to 1110.10 or any
downgrades from 1120 to 1110.10 would wipe all the persisted data &
starts clean.

Change-Id: I698556a1b9781d7d869daf18348bdcf6ec82d6a7
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>